requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
azure-ai-documentintelligence==1.0.0b4
azure-core==1.36.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.120.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jupyter_client==8.6.3
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
openai==2.6.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
